Check and requets for read permission on Android.
This is required on Android 10+.

Otherwise, no images or videos will be shown.

Also confirmed to be working on Android 5, which doesn't show the request if the permission is already added to the app's AndroidManifest file.

The check and request method was pulled directly from cameraroll's README, with the exception of just checking the READ permission instead of the WRITE permission:

https://github.com/react-native-cameraroll/react-native-cameraroll#permissions.