Fetch `playableDuration` by default.
In additional to providing the duration of videos, it will also remove thumbnails where files do not exist.

For example, `/storage/emulated/0/ddmsrec.mp4` was being shown as a thumbnail even though this does not exist. The result was an empty thumbnail and an inability/error when trying to select that file. This fixes that.

For context, I believe this phantom file was created during video recording from Android Studio, but I'm not positive.