Sort photos by date.
For some reason, `CameraRoll` doesn’t return photos in chronological order. See https://github.com/react-native-cameraroll/react-native-cameraroll/issues/45.

Also upgraded `react-native-cameraroll` from 4.0.1 to 4.0.4. 4.0.3 release notes say that it fixes an issue where the `timestamp` of the image was returned as `0`. I was seeing this on the Tab A (SM-T290) with Android 10.